chore: rename test files after refactor
