+++ lib/modules/pool/actions/add-liquidity/handlers/UnbalancedAddLiquidityV3.handler.ts
@@ -6,13 +6,13 @@ import { BaseUnbalancedAddLiquidityHandler } from './BaseUnbalancedAddLiquidity.
 import { constructBaseBuildCallInput } from './add-liquidity.utils'
 
 /**
- * UnbalancedAddLiquidityHandlerV3 is a handler that implements the
+ * UnbalancedAddLiquidityV3Handler is a handler that implements the
  * AddLiquidityHandler interface for unbalanced V3 adds, e.g. where the user
  * specifies the token amounts in. It uses the Balancer SDK to implement it's
  * methods. It also handles the case where one of the input tokens is the native
  * asset instead of the wrapped native asset.
  */
-export class UnbalancedAddLiquidityHandlerV3 extends BaseUnbalancedAddLiquidityHandler {
+export class UnbalancedAddLiquidityV3Handler extends BaseUnbalancedAddLiquidityHandler {
   public async buildCallData({
     humanAmountsIn,
     slippagePercent,
