+++ lib/modules/pool/actions/add-liquidity/handlers/selectAddLiquidityHandler.ts
@@ -7,7 +7,7 @@ import { NestedAddLiquidityHandler } from './NestedAddLiquidity.handler'
 import { requiresProportionalInput, supportsNestedActions } from '../../LiquidityActionHelpers'
 import { ProportionalAddLiquidityHandler } from './ProportionalAddLiquidity.handler'
 import { isV3Pool } from '../../../pool.helpers'
-import { UnbalancedAddLiquidityHandlerV3 } from './UnbalancedAddLiquidityV3.handler'
+import { UnbalancedAddLiquidityV3Handler as UnbalancedAddLiquidityV3Handler } from './UnbalancedAddLiquidityV3.handler'
 
 export function selectAddLiquidityHandler(pool: Pool): AddLiquidityHandler {
   // This is just an example to illustrate how edge-case handlers would receive different inputs but return a common contract
@@ -24,7 +24,7 @@ export function selectAddLiquidityHandler(pool: Pool): AddLiquidityHandler {
     return new NestedAddLiquidityHandler(pool)
   }
 
-  if (isV3Pool(pool)) return new UnbalancedAddLiquidityHandlerV3(pool)
+  if (isV3Pool(pool)) return new UnbalancedAddLiquidityV3Handler(pool)
 
   return new UnbalancedAddLiquidityV2Handler(pool)
 }
